refactor: move Netlify functions inside directories
